Rename package to `streamflow_postgresql`
This commit refactors the package name into `streamflow_postgresql` to
avoid package name conflicts.
