+++ README.md
@@ -11,7 +11,7 @@ pip install streamflow-postgresql
 If everything worked correctly, whenever a workflow execution start the following message should be printed in the log:
 
 ```bash
-Successfully registered plugin postgresql.plugin.PostgreSQLStreamFlowPlugin
+Successfully registered plugin streamflow_postgresql.plugin.PostgreSQLStreamFlowPlugin
 ```
 
 ## Usage
