+++ tox.ini
@@ -27,7 +27,7 @@ setenv =
   py3.{8,9,10,11}-unit: LC_ALL = C.UTF-8
 
 [testenv:bandit]
-commands = bandit -r postgresql
+commands = bandit -r streamflow_postgresql
 deps =
   -rrequirements.txt
   -rbandit-requirements.txt
